Moved Create Command to right namespace
