+++ HubOfChess.Application/PostLikes/Commands/CreatePostLike/CreatePostLikeCommand.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace HubOfChess.Application.PostLikes.Commands
+namespace HubOfChess.Application.PostLikes.Commands.CreatePostLike
 {
     public record CreatePostLikeCommand(Guid UserId, Guid PostId) : IRequest;
 }
